Initial reports content.
