+++ webapp/package.json
@@ -20,6 +20,7 @@
     "@angular/platform-browser": "^10.0.14",
     "@angular/platform-browser-dynamic": "^10.0.14",
     "@angular/router": "^10.0.14",
+    "@ngrx/effects": "^10.0.0",
     "@ngrx/store": "^10.0.0",
     "@ngx-translate/core": "^13.0.0",
     "bikram-sambat": "^1.6.0",
@@ -39,7 +40,7 @@
     "ng-redux": "^4.4.1",
     "ngrx-store-logger": "^0.2.4",
     "ngx-bootstrap": "^6.1.0",
-    "ngx-cookie": "^5.0.0",
+    "ngx-cookie-service": "^10.0.1",
     "object-path": "^0.11.4",
     "odk-xform-compact-record-representation-for-sms": "^1.0.1",
     "openrosa-xpath-evaluator": "^1.5.1",
