+++ webapp/src/ts/actions/reports.ts
@@ -2,6 +2,8 @@ import { Store, createAction } from '@ngrx/store';
 import { createSingleValueAction } from './actionUtils';
 
 export const Actions = {
+  selectReport: createSingleValueAction('SELECT_REPORT', 'id'),
+  setSelected: createSingleValueAction('SET_SELECTED_REPORT', 'selected'),
   addSelectedReport: createSingleValueAction('ADD_SELECTED_REPORT', 'selected'),
   removeSelectedReport: createSingleValueAction('REMOVE_SELECTED_REPORT', 'id'),
   setSelectedReports: createSingleValueAction('SET_SELECTED_REPORTS', 'selected'),
@@ -16,6 +18,10 @@ export class ReportsActions {
 
   addSelectedReport(selected) {
     return this.store.dispatch(Actions.addSelectedReport(selected));
+  }
+
+  selectReport(id) {
+    return this.store.dispatch(Actions.selectReport(id));
   }
 
   removeSelectedReport(id) {
@@ -42,6 +48,74 @@ export class ReportsActions {
   resetReportsList() {
     return this.store.dispatch(Actions.resetReportsList());
   }
+
+  /*setSelected(model) {
+    this.store.
+
+    dispatch(function(dispatch, getState) {
+      const selectMode = Selectors.getSelectMode(getState());
+      const selectedReports = Selectors.getSelectedReports(getState());
+      let refreshing = true;
+      if (selectMode) {
+        const existing = _.find(selectedReports, { _id: model.doc._id });
+        if (existing) {
+          Object.assign(existing, model);
+        } else {
+          model.expanded = false;
+          addSelectedReport(model);
+        }
+      } else {
+        if (LiveList.reports.initialised()) {
+          LiveList.reports.setSelected(model.doc && model.doc._id);
+          LiveList['report-search'].setSelected(model.doc && model.doc._id);
+        }
+        refreshing =
+          model.doc &&
+          selectedReports.length &&
+          selectedReports[0]._id === model.doc._id;
+        if (!refreshing) {
+          setVerifyingReport(false);
+        }
+
+        model.expanded = true;
+        setSelectedReports([model]);
+        setTitle(model);
+
+        const listModel = LiveList.reports.getList().find(item => item._id === model._id);
+        if (listModel && !listModel.read) {
+          const unreadCount = Selectors.getUnreadCount(getState());
+          globalActions.updateUnreadCount({ report: unreadCount.report - 1 });
+          listModel.read = true;
+          LiveList.reports.update(listModel);
+          LiveList['report-search'].update(listModel);
+          MarkRead([model.doc]).catch(err => $log.error('Error marking read', err));
+        }
+      }
+      setRightActionBar();
+      globalActions.settingSelected(refreshing);
+    });
+  }
+
+  selectReport(id, { silent=false }={}) {
+    if (!id) {
+      return Promise.resolve();
+    }
+    if (!silent) {
+      this.store.dispatch(GlobalActions.setLoadingContent(true));
+      this.store.dispatch(GlobalActions.setShowContent(true));
+    }
+    return this.reportViewModelGeneratorService
+      .get(id)
+      .then(model => {
+        if (model) {
+          setSelected(model);
+        }
+      })
+      .catch(err => {
+        globalActions.unsetSelected();
+        $log.error('Error selecting report', err);
+      });
+  };*/
 
 }
 /*
@@ -142,50 +216,7 @@ angular.module('inboxServices').factory('ReportsActions',
         });
       }
 
-      function setSelected(model) {
-        dispatch(function(dispatch, getState) {
-          const selectMode = Selectors.getSelectMode(getState());
-          const selectedReports = Selectors.getSelectedReports(getState());
-          let refreshing = true;
-          if (selectMode) {
-            const existing = _.find(selectedReports, { _id: model.doc._id });
-            if (existing) {
-              Object.assign(existing, model);
-            } else {
-              model.expanded = false;
-              addSelectedReport(model);
-            }
-          } else {
-            if (LiveList.reports.initialised()) {
-              LiveList.reports.setSelected(model.doc && model.doc._id);
-              LiveList['report-search'].setSelected(model.doc && model.doc._id);
-            }
-            refreshing =
-              model.doc &&
-              selectedReports.length &&
-              selectedReports[0]._id === model.doc._id;
-            if (!refreshing) {
-              setVerifyingReport(false);
-            }
 
-            model.expanded = true;
-            setSelectedReports([model]);
-            setTitle(model);
-
-            const listModel = LiveList.reports.getList().find(item => item._id === model._id);
-            if (listModel && !listModel.read) {
-              const unreadCount = Selectors.getUnreadCount(getState());
-              globalActions.updateUnreadCount({ report: unreadCount.report - 1 });
-              listModel.read = true;
-              LiveList.reports.update(listModel);
-              LiveList['report-search'].update(listModel);
-              MarkRead([model.doc]).catch(err => $log.error('Error marking read', err));
-            }
-          }
-          setRightActionBar();
-          globalActions.settingSelected(refreshing);
-        });
-      }
 
       const selectReport = (id, { silent=false }={}) => {
         if (!id) {
